feature: SRYD-416

+++ glance/glance/cluster/services/cluster-curd.service.js
@@ -8,7 +8,7 @@
 
 
     /* @ngInject */
-    function clusterCurd(clusterBackend, confirmModal, $state, Notification) {
+    function clusterCurd(clusterBackend, confirmModal, $state, Notification, appservice, $mdDialog) {
         return {
             listClusterLables: listClusterLables,
             deleteCluster: deleteCluster,
@@ -32,17 +32,36 @@
                 });
         }
 
-        function deleteCluster(clusterId, ev) {
-            var content = '删除集群将在您的主机上一并清除：数人云管理组件，通过数人云下发部署的应用及其未持久化的数据';
-            confirmModal.open('您确定要删除集群吗？', ev, content)
-                .then(function () {
-                    clusterBackend.deleteCluster(clusterId)
-                        .then(function (data) {
-                            Notification.success('集群删除成功');
-                            $state.go('cluster.list', null, {reload: true});
-                        });
-                })
-
+        function deleteCluster(clusterId, ev, nodeTotal) {
+            appservice.listClusterApps({}, clusterId).then(function(data) {
+                var content;
+                if (data.Count > 0) {
+                    content = '集群中还有' + data.Count + '个应用而无法删除,请先删除应用';
+                    var confirm = $mdDialog.confirm()
+                        .clickOutsideToClose(true)
+                        .title('无法删除集群')
+                        .targetEvent(ev)
+                        .ok('去看看应用')
+                        .cancel('取消').textContent(content);
+                    var dialog = $mdDialog.show(confirm);
+                    dialog.then(function() {
+                        $state.go('app');
+                    });
+                } else {
+                    if(nodeTotal > 0) {
+                        content = '集群中还有' + nodeTotal + '台主机，删除集群将同时清除这些主机上的数人云各组件。<br>'
+                            + '此操作不可恢复，请慎重选择！'
+                    }
+                    confirmModal.open('您确定要删除集群吗？', ev, content)
+                        .then(function () {
+                            clusterBackend.deleteCluster(clusterId)
+                                .then(function (data) {
+                                    Notification.success('集群删除成功');
+                                    $state.go('cluster.list', null, {reload: true});
+                                });
+                        })
+                }
+            });
         }
 
         function deletNodes(clusterId, selectNodes, ev) {
